+++ plugin/src/main/java/com/stratio/cassandra/lucene/search/SearchBuilders.java
@@ -15,6 +15,7 @@
  */
 package com.stratio.cassandra.lucene.search;
 
+import com.stratio.cassandra.lucene.common.GeoShape;
 import com.stratio.cassandra.lucene.search.condition.builder.*;
 import com.stratio.cassandra.lucene.search.sort.builder.GeoDistanceSortFieldBuilder;
 import com.stratio.cassandra.lucene.search.sort.builder.SimpleSortFieldBuilder;
@@ -245,11 +246,24 @@ public final class SearchBuilders {
      * /** Constructor receiving the name of the field and the shape.
      *
      * @param field the name of the field
+     * @param shape the shape
+     * @return a new geo shape condition builder
+     */
+    public static GeoShapeConditionBuilder geoShape(String field, GeoShape shape) {
+        return new GeoShapeConditionBuilder(field, shape);
+    }
+
+    /**
+     * Returns a new {@link GeoShapeConditionBuilder} with the specified field reference point.
+     *
+     * /** Constructor receiving the name of the field and the shape.
+     *
+     * @param field the name of the field
      * @param shape the shape in <a href="http://en.wikipedia.org/wiki/Well-known_text"> WKT</a> format
      * @return a new geo shape condition builder
      */
     public static GeoShapeConditionBuilder geoShape(String field, String shape) {
-        return new GeoShapeConditionBuilder(field, shape);
+        return new GeoShapeConditionBuilder(field, new GeoShape.WKT(shape));
     }
 
     /**