+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/varia/BoundStatementWithSortedKQuery.java
@@ -15,6 +15,7 @@
  */
 package com.stratio.cassandra.lucene.testsAT.varia;
 
+import com.stratio.cassandra.lucene.builder.Builder;
 import com.stratio.cassandra.lucene.testsAT.search.AbstractSearchAT;
 import com.stratio.cassandra.lucene.testsAT.util.CassandraUtilsSelect;
 import org.junit.Test;
@@ -91,28 +92,28 @@ public class BoundStatementWithSortedKQuery extends AbstractSearchAT {
     @Test
     public void sortWithGeoDistanceFilterNotReversed() {
         utils.filter(geoDistance("geo_point", 40.442163, -3.784519, "10000km"))
-             .sort(geoDistanceField("geo_point", 40.442163, -3.784519).reverse(false))
+             .sort(Builder.geoDistance("geo_point", 40.442163, -3.784519).reverse(false))
              .checkOrderedColumns("integer_1", -1, -2, -3, -4, -5);
     }
 
     @Test
     public void sortWithGeoDistanceQueryNotReversed() {
         utils.query(geoDistance("geo_point", 40.442163, -3.784519, "10000km"))
-             .sort(geoDistanceField("geo_point", 40.442163, -3.784519).reverse(false))
+             .sort(Builder.geoDistance("geo_point", 40.442163, -3.784519).reverse(false))
              .checkOrderedColumns("integer_1", -1, -2, -3, -4, -5);
     }
 
     @Test
     public void sortWithGeoDistanceFilterReversed() {
         utils.filter(geoDistance("geo_point", 40.442163, -3.784519, "10000km"))
-             .sort(geoDistanceField("geo_point", 40.442163, -3.784519).reverse(true))
+             .sort(Builder.geoDistance("geo_point", 40.442163, -3.784519).reverse(true))
              .checkOrderedColumns("integer_1", -5, -4, -3, -2, -1);
     }
 
     @Test
     public void sortWithGeoDistanceQueryReversed() {
         utils.query(geoDistance("geo_point", 40.442163, -3.784519, "10000km"))
-             .sort(geoDistanceField("geo_point", 40.442163, -3.784519).reverse(true))
+             .sort(Builder.geoDistance("geo_point", 40.442163, -3.784519).reverse(true))
              .checkOrderedColumns("integer_1", -5, -4, -3, -2, -1);
     }
 }