+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 3.0.8.2 (Upcoming)
 
+* Allow associativity in search-time geospatial transformations
+* Simplify naming of builder static methods for creating geospatial transformations
+
 ## 3.0.8.1 (August 02, 2016)
 
 * Upgrade query builder JSON serializer to Jackson 2.8.0